mGerrit: Remove child ListView from parent ListView (CardsUI)
Embedding ListViews in ListViews causes the rendering to of the
child ListView to only maintain its layout_heigth before its
Adapter is set. Then when the child ListView is populated by
the Adapter we don't see the entire list.

This method generates a view for each Reviewer then attaches that
View to the actual card not a ListView container.

Change-Id: I48188c78429de88785257525b47fedc896d00196