Master-detail flow conversion
Converted PatchSetViewer into a fragment
Added new master-detail layout, with the tablet layout split in half
 (half screen for each fragment)
PatchSetViewerFragement maintains the changeID itself.
Included the fragment as a seperate layout file.
PatchSetViewerFragment: Disable caching. Similarly to
 GerritControllerActivity, this caching mechanism is flawed
Save and restore selected change (for tablet layouts) in database.
Use card onClick listener instead of a small icon on tablets.
Store and retreive the legacy change number in the database so we
 can obtain the change URL in the changes list.

PS2:  This fragment is a container for the viewpager.
Use a PagerTabStrip instead of the action bar for the tabs. This looks a bit
 different.
Added missing SelectedChange database table class.
PS3: Fix tablet layout issues.

Change-Id: If76feb28b904d895a3bc5480b527960ba321a465