Fix NullPointerException seemingly cause by SwipeRefreshLayout
The Exception was because previously the SwipeRefreshLayout reference
was set when we created the underlying Activity that loaded
CardsFragment.  Since CardsFragment is a super class for ReviewTab,
MergeTab and AbandonTab the previous behavior was problematic as
the refrence would only get set for one Fragment (ReviewTab). The
new behavior is to gain the reference everytime we create a subclass
of CardsFragment.  Since all subclasses of CardsFragment will now
have the correct reference to the SwipeRefreshLayout the library
behaves as expected.

Change-Id: Ibf8015294d4ad38d0dc90d8aed19d4ea4da08a9c