Fix before and after searching logic
The logic for before and after searching was not consistent with the one on the Gerrit web interface.
This fixes the logic so that after will search for more recent changes and before will search for older
changes than the date given.
Don't broadcast the change loading finished message if not in tablet mode as it is only used when
by the patch set viewer panel in the master-detail layout.
Update Gradle tools version with latest Android studio version

Change-Id: I6f71635cab2fbcb94b468569f09d47984918f9a2