Request username/password when request requires authentication
This adds support for retrieving changes when the Gerrit requires authentication.
We will bring up the sign in screen if it has not already been started, request for a username and password, then close the sign in screen so changes are visible
Also fix an error when a user does not have an email address (e.g. the Gerrit user)
Authentication handling has been added for fetching changes and projects.
Reduce no subscriber event log spam from EventBus

Change-Id: Ic657dc52e651c10fb36ca046f1a15d1354572527
TODO: Add authentication handling for retrieving the server version, change diffs and images
TODO: Show a different message on the sign in activity when sign in is required