Add ConfigEndpoints class
This completes hooking up authentication to VersionProcessor .
However, as EventBus only retains the last sticky event and we are queuing up multiple requests,
an ErrorDuringConnection event is lost so we get the server version, but don't re-query for changes.
TODO: Add an event queue for them to be consumed by the SigninActivity and then removed once processed.

Change-Id: Ibd1c09afb9cfa98f3c1bab23d515ed5d21d710a7