Set all changes as unstarred when logging out
As the user may log out and log in as someone else, set all their changes (locally) as unstarred when logging out.
This does not send a request out, so their changes will appear as starred when they log back in.
Use a default Gravatar image and always show the profile picture on the commit cards to prevent broken layouts.

Change-Id: Ic73ff312c67efefd14f00239c7d407b09265ddf0