Fix - Quick fix for Gerrits not loading newer changes
We only request the server version once, so if that Gerrit is updated
we do not record the updated version. This is a problem when they were
updated to a version which removed the sortkey (v2.8.1) as fetching
newer changes will not work.

This is a temporary fix, selecting the option to refresh from the menu
will fetch the latest server version. Then a pull to refresh will get
the latest changes

Change-Id: I0a5a820c5feb96934edb7baa15e41261e9d06d45