Complete deprecation of AOKP changelog functionality
The AOKP changelog has been broken for a while and not been maintained for even
longer. Remove it as it is not returning. Gerrit does not have any way of tracking what
changes where included in a build (the build needs to contain information relating to
what commits were included).

Update grade plugin
Add initial branch search definition

Change-Id: I42e71756bccad913710068ceb39aeb9b968a687a