Save and restore diff text state when orientation changes
Don't make a new request to fetch the data from the cache if the diff has already loaded
and it is text.
This fixes one case where we were making unnecessary requests, but we also don't need
to make duplicate requests when the orientation changes while the diff is loading.
Another attempt to fix possible crash where the action bar is not completely removed

refs #48

Change-Id: I454e619c1a375d5cafb71558172c71b1bdbf1e4f