BUGFIX - Toggle refine search card visibility rather than adding/removing
setHeaderView cannot be called after setting the adapter in Android KK and below.,
as a result, we need to toggle the visibility of the card which will show/hide it without
actually removing it.
This also fixes the issue where the date was not displayed for the first item in the list.

refs #47

Change-Id: I16dfbeaaa29ef90d4e64cb9f24c636ba2c6ec2ca