Automatic night mode theme option
Automatically switch the theme based on the time of day. The support library
will use the location if available, but we don't want to request that so the auto
option may not be very precise.

Re-structure themes resources
Upgrade build tools version
Switch to uwolfer version of Gerrit REST library (android fork) now it is up-to-date
Fix Crashlytics crashing on dev
More Crashlytics logging
Fix possible NPE in when inserting labels if no default value is set

Change-Id: I08a270becd60df54f2e420e9e24701b9d6dd5600